Rename caf::flow::{item_publisher => multicaster}
Since we are using the term "publisher" to refer to objects that publish
items asynchronously now, the name `item_publisher` is likely to cause
confusion for new users. The new name also better reflects the purpose.
