+++ libcaf_core/CMakeLists.txt
@@ -287,8 +287,8 @@ caf_add_component(
     flow.flat_map
     flow.for_each
     flow.generation
-    flow.item_publisher
     flow.mixed
+    flow.multicaster
     flow.observe_on
     flow.op.buffer
     flow.op.cell
